Update program.cpp
The first ball of each level is now held by the paddle until the player presses player 1's button 1 to avoid instant death, and added on screen prompt to tell the player how to release the ball. Left the debugging cheats enabled but changed the key-binds to keys not usable by the arcade machine keyboard emulator, which should make cheating impossible with the arcade machines. Also fixed the initials entry screen retaining the selection of the right initial after submitting.